Add common test-cases repo and integrate RLP tests
Incurring a dependency on aeson; only in the test suite for now. I was
trying to avoid using the typeclass style for some reason I can't
remember, but I think it's nicer in the end.

Happy with how simple the switch to Integral n vs. Int was in
{encode,decode}Int; that would suck in all other languages I've used,
especially since it's polymorphic in the return type.

Unlike fixing the QuickCheck, this did not uncover any bugs.