Refactor MapDB into the library proper
This implementation is actually somewhat nicer, since it hides the state
monad and exposes only the DB interface and a runner.

This doesn't allow you to set the initial state, or recover the state
from runState. I think this is fine -- impossible to pass in an invalid
state, and if you want to inspect the state you can do it from within
the monad -- but I may need to add this later if new requirements crop
up.