Another nonsensical main method
This is a tiny TCP server that accepts connections from multiple hosts,
first printing the socket address and then printing all the received
input until the connection is closed.

This uses pipes only because I'm hoping to use pipes-attoparsec later
on, so I wanted to kick the tires. Pipes scares me a little but so far
the obvious things have been working just fine.