Rework the whole DB code into something more useful
A bit of a drag, but I like the API much better now.

There's a single abstraction for map-style operations, as bundled up in
the DB type. This means that the Patricia Trie and the underlying
database can implement the same interface, which is handy. (And both can
be swapped with the map-based version for testing.) Also, the
state-monad-based Trie eliminates the manual passing-around of the
trie's root hash. (Except for the internal implementation, where that's
exactly what you want.)