Sidetrack - rewrite some of the RLP helpers
I was writing some RLP conversions that really shouldn't have needed to
be done manually, and it's way more fun to do this generically.

We now wrap the to- and from- conversion functions in a struct. This
makes it easier to operate on both consistently, like by adding 'tags'.
The code can now derive two different representations -- one general one
where constructors translate to products (without special-casing the
single-element constructor) and another that only compiles for
single-element constructors (or sums thereof) and keeps everything
unwrapped. This makes things more explicit when we're relying on
something unusual, but without a lot of boilerplate. Alright!