Fix an issue with ConvertProduct, plus some renaming
Given a datatype like so:

    data Foo = Bar Int | Baz Int Int

ConvertProduct should deserialize as Baz if the list has two elements.
However, the previous version was deserializing as Bar with a
single-element leftover; and since we disallow leftovers at the top
level, deserialization was failing.

The fix is to disallow leftovers farther down, inside the sum of the
sum-of-products. This fails sooner, so deserialization can fall down the
other branch.