Push normalization into the insert function, and remove some recursion
Instead of renormalizing the trie each time, which grows increasingly
wasteful as the tree enlarges, we should normalize locally as we make
changes.

I suspect this is still not minimal, but at least it's *local* --
the work on each insert will be proportional to the depth of the tree,
not the entire size.